requirements.txt
aiogram==2.20
requests==2.28.0
aioschedule~=0.5.2
google-api-core==1.31.6
google-auth==1.35.0
google-cloud-dialogflow==2.14.1
googleapis-common-protos==1.56.2
